fix: #1287 migrate graphql-server to serverless offline
