--- packages/graphql-server/nodemon.json
@@ -1,6 +1,0 @@
-{
-  "watch": ["src"],
-  "ext": "ts,ejs,graphql",
-  "ignore": ["src/**/*.test.ts", "src/schema.graphql"],
-  "exec": "ts-node -r tsconfig-paths/register -P tsconfig.json ./src/app.ts"
-}
