+++ packages/graphql-server/package.json
@@ -1,6 +1,6 @@
 {
   "name": "graphql-server",
-  "version": "0.0.0",
+  "version": "0.0.1",
   "private": true,
   "description": "GraphQL Server",
   "repository": "https://github.com/reapit/foundations.git",
@@ -11,9 +11,10 @@
     "build:schema": "node ./scripts/build-schema.js && prettier src/schema.graphql --write",
     "lint": "concurrently \"tsc --noEmit\" \"eslint --cache --ext=ts,js src\"",
     "lint:fix": "prettier \"src/**/*.{graphql,gql}\" --write && eslint --cache --ext=ts,js src --fix",
+    "release:dev": "yarn build:schema && serverless deploy --stage dev",
     "release:development": "yarn build:schema && serverless deploy --stage dev",
     "release:production": "yarn build:schema && serverless deploy --stage prod",
-    "start:dev": "yarn build:schema && cross-env NODE_ENV=development nodemon",
+    "start:dev": "yarn build:schema && serverless offline --stage local",
     "start:prod": "node ./dist/app.js",
     "test:ci": "cross-env TZ=UTC jest --ci --colors --coverage --silent --forceExit",
     "test:dev": "cross-env TZ=UTC jest --watch --verbose",
@@ -23,25 +24,19 @@
     "prepublish": "npm run snyk-protect"
   },
   "dependencies": {
-    "apollo-server": "^2.9.16",
     "apollo-server-lambda": "^2.10.1",
-    "async": "^3.2.0",
     "axios": "^0.19.2",
-    "color-convert": "^2.0.1",
-    "color-name": "^1.1.4",
-    "colors": "^1.4.0",
+    "cookie": "0.4.1",
     "dataloader": "^2.0.0",
-    "express": "^4.17.1",
     "graphql": "^14.5.8",
     "graphql-depth-limit": "^1.1.0",
     "graphql-import": "^0.7.1",
     "graphql-type-json": "^0.3.1",
-    "is-stream": "^2.0.0",
+    "https-proxy-agent": "5.0.0",
     "lodash.merge": "^4.6.2",
     "merge-graphql-schemas": "^1.7.6",
-    "ms": "^2.1.2",
-    "readable-stream": "^3.6.0",
-    "stack-trace": "^0.0.10",
+    "query-string": "^6.13.0",
+    "tslib": "2.0.0",
     "uuid": "^3.3.3",
     "winston": "^3.2.1",
     "snyk": "^1.335.0",
@@ -62,13 +57,8 @@
     "bufferutil": "^4.0.1",
     "copy-webpack-plugin": "^5.1.1",
     "jest": "^25.1.0",
-    "json-loader": "^0.5.7",
-    "nodemon": "^2.0.2",
-    "serverless-webpack": "^5.3.1",
-    "ts-node": "^8.5.4",
-    "tsconfig-paths": "^3.9.0",
-    "tsdx": "^0.12.3",
-    "utf-8-validate": "^5.0.2",
+    "serverless-offline": "^6.1.4",
+    "serverless-webpack": "^5.3.2",
     "webpack-node-externals": "^1.7.2"
   },
   "snyk": true
