+++ packages/graphql-server/src/resolvers/appointments/appointments.d.ts
@@ -1,4 +1,4 @@
-import { AuthenticationError, UserInputError } from 'apollo-server'
+import { AuthenticationError, UserInputError } from 'apollo-server-lambda'
 import {
   AppointmentModel,
   PagedResultAppointmentModel_,
