+++ packages/graphql-server/src/resolvers/configurations/services.ts
@@ -1,5 +1,5 @@
 import logger from '../../logger'
-import { ServerContext } from '../../app'
+import { ServerContext } from '../../index'
 import {
   GetConfigurationByTypeAndIdReturn,
   GetConfigurationsByTypeReturn,
