+++ packages/graphql-server/src/resolvers/contacts/api.ts
@@ -1,6 +1,6 @@
 import qs from 'query-string'
 import logger from '../../logger'
-import { ServerContext } from '../../app'
+import { ServerContext } from '../../index'
 import {
   GetContactByIdArgs,
   CreateContactArgs,
