+++ packages/graphql-server/src/resolvers/contacts/contacts.d.ts
@@ -1,4 +1,4 @@
-import { AuthenticationError, UserInputError } from 'apollo-server'
+import { AuthenticationError, UserInputError } from 'apollo-server-lambda'
 import { ContactModel, PagedResultContactModel_, CreateContactModel, UpdateContactModel } from '../../types'
 
 export type CreateContactArgs = CreateContactModel
