+++ packages/graphql-server/src/resolvers/identity-checks/identity-checks.d.ts
@@ -1,4 +1,4 @@
-import { AuthenticationError, UserInputError } from 'apollo-server'
+import { AuthenticationError, UserInputError } from 'apollo-server-lambda'
 import {
   IdentityCheckModel,
   PagedResultIdentityCheckModel_,
