+++ packages/graphql-server/src/resolvers/negotiators/resolvers.ts
@@ -2,7 +2,7 @@ import negotiatorServices from './services'
 import { checkPermission } from '../../utils/check-permission'
 import logger from '../../logger'
 import errors from '../../errors'
-import { ServerContext } from '../../app'
+import { ServerContext } from '../../index'
 import {
   GetNegotiatorByIdArgs,
   CreateNegotiatorArgs,
