+++ packages/graphql-server/src/resolvers/negotiators/services.ts
@@ -1,5 +1,5 @@
 import logger from '../../logger'
-import { ServerContext } from '../../app'
+import { ServerContext } from '../../index'
 import {
   GetNegotiatorByIdArgs,
   CreateNegotiatorArgs,
