+++ packages/graphql-server/src/utils/check-permission.ts
@@ -1,5 +1,5 @@
 import logger from '../logger'
-import { ServerContext } from '../app'
+import { ServerContext } from '../index'
 
 export const checkPermission = (context: ServerContext) => {
   const traceId = context.traceId
